stim/requirements.txt
annotated-types~=0.6.0
anyio~=3.7.1
exceptiongroup~=1.1.3
fastapi~=0.103.2
idna~=3.4
pydantic~=2.4.2
pydantic_core~=2.10.1
sniffio~=1.3.0
starlette~=0.27.0
typing_extensions~=4.8.0
black~=23.9.1
uvicorn~=0.23.2
discord.py~=2.3.2
python-dotenv~=1.0.0
langchain~=0.0.310
openai~=0.28.1
